# [solidity] improve overall performance