--- a/regression/esbmc-solidity/call_1/test.desc
+++ b/regression/esbmc-solidity/call_1/test.desc
@@ -1,4 +1,4 @@
 CORE
 contract.solast
---sol contract.sol --k-induction --no-standard-checks --unbound
+--sol contract.sol --k-induction --no-standard-checks --unbound --contract Bank
 ^VERIFICATION FAILED$