--- a/regression/esbmc-solidity/call_5/test.desc
+++ b/regression/esbmc-solidity/call_5/test.desc
@@ -1,4 +1,4 @@
 CORE
 contract.solast
---sol contract.sol --no-standard-checks --unwind 3 --no-unwinding-assertions
+--sol contract.sol --no-standard-checks --unwind 3 --no-unwinding-assertions --contract Reproduction
 ^VERIFICATION SUCCESSFUL$