--- a/regression/esbmc-solidity/doftcoin_1/test.desc
+++ b/regression/esbmc-solidity/doftcoin_1/test.desc
@@ -1,4 +1,4 @@
 CORE
 contract.solast
---sol contract.sol --contract Doftcoin --incremental-bmc --unsigned-overflow-check --unbound
+--sol contract.sol --contract Doftcoin --incremental-bmc --unsigned-overflow-check --unbound --function mintToken
 ^VERIFICATION FAILED$