--- a/regression/esbmc-solidity/doftcoin_1/test.desc
+++ b/regression/esbmc-solidity/doftcoin_1/test.desc
@@ -1,0 +1,4 @@
+CORE
+contract.solast
+--sol contract.sol --contract Doftcoin --incremental-bmc --unsigned-overflow-check --unbound
+^VERIFICATION FAILED$