--- a/regression/esbmc-solidity/doftcoin_2/test.desc
+++ b/regression/esbmc-solidity/doftcoin_2/test.desc
@@ -1,4 +1,0 @@
-CORE
-contract.solast
---sol contract.sol --incremental-bmc --no-standard-checks --contract Reproduction --unsigned-overflow-check
-^VERIFICATION FAILED$