--- a/regression/esbmc-solidity/doftcoin_2/test.desc
+++ b/regression/esbmc-solidity/doftcoin_2/test.desc
@@ -1,0 +1,4 @@
+CORE
+contract.solast
+--sol contract.sol --contract Reproduction --incremental-bmc --no-standard-checks --unsigned-overflow-check
+^VERIFICATION FAILED$