--- a/regression/esbmc-solidity/mapping_13/test.desc
+++ b/regression/esbmc-solidity/mapping_13/test.desc
@@ -1,4 +1,4 @@
 KNOWNBUG
 contract.solast
 --sol contract.sol --unwind 3 --no-unwinding-assertions
-^VERIFICATION FAILED$
+^VERIFICATION SUCCESSFUL$