--- a/regression/esbmc-solidity/mapping_16/test.desc
+++ b/regression/esbmc-solidity/mapping_16/test.desc
@@ -1,0 +1,4 @@
+KNOWNBUG
+contract.solast
+--sol contract.sol --k-induction --contract Derived 
+^VERIFICATION FAILED$