--- a/regression/esbmc-solidity/mul_cnt_ver_1/test.desc
+++ b/regression/esbmc-solidity/mul_cnt_ver_1/test.desc
@@ -1,4 +1,4 @@
 CORE
 contract.solast
---sol contract.sol --multi-property --incremental-bmc
+--sol contract.sol --multi-property --k-induction --no-standard-checks
 ^VERIFICATION FAILED$