--- a/regression/esbmc-solidity/reentrance_13/test.desc
+++ b/regression/esbmc-solidity/reentrance_13/test.desc
@@ -1,0 +1,4 @@
+CORE
+contract.solast
+--sol contract.sol --contract Reproduction --no-standard-checks --incremental-bmc
+^VERIFICATION FAILED$