```sh
# tested on reentrance_1
# before
Over 5 mins
# after
esbmc --sol example.sol example.solast --incremental-bmc --no-standard-checks  1.90s user 0.19s system 64% cpu 3.255 total

# tested on reentrance_2
# before
Over 5 mins
# after
esbmc --sol example.sol example.solast --contract Reproduction     4.13s user 0.20s system 80% cpu 5.384 total
```

![image](https://github.com/user-attachments/assets/f387cf4a-b3a6-4163-8aa7-d8594ce6bc95)

now the whole regression test can be finished in 120 second (-j 8)
---

This PR also contains some bug fixing, including library API (_max/_min), address binding in trusted mode